refactor(search): se separo la logica en un custom hook
