+++ src/pages/repositories/index.js
@@ -1,47 +1,20 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { Text } from "components/Text";
 import { Layout } from "components/Layout";
 import { SearchBar } from "components/SearchBar";
 import { Card } from "components/Card";
-import { useForm } from "react-hook-form";
-import { useQuery } from "react-query";
 import { getRepositorie } from "services";
 import { Loader } from "components/Loader";
-import useDebounce from "utils";
+import { useSearch } from "hooks";
 
 const ReposotiriesPage = () => {
-
-  const {
-    watch,
-    register,
-    handleSubmit,
-  } = useForm();
-    
-  const name = watch("search");
-
-  const debouncedSearch = useDebounce(name, 1000);
-
-  const { data, refetch, isLoading } = useQuery(
-    ['repositories', debouncedSearch],
-    getRepositorie,
-    {
-      enabled: !!debouncedSearch,
-      refetchOnWindowFocus: false,
-    }
-  );
-
-  useEffect(() => {
-    if(debouncedSearch) {
-      refetch();
-    }
-  }, [debouncedSearch]);
-
+  const { data, isLoading, register, name } = useSearch(getRepositorie, "repositories");
   return(  
     <Layout path="repositories">
       <Text variant="big" type="h1">
         Busca un repositorio 
       </Text>
-      <SearchBar action={handleSubmit((data) => refetch())} register={register}/>
+      <SearchBar register={register}/>
       { isLoading && <Loader />}
       { (name && data?.length === 0) && `No hay resultados para la busqueda: ${name}` }
       {
